main_DL/requirements.txt
# Deep Learning branch dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
imbalanced-learn>=0.10.0
tensorflow-cpu>=2.12.0
